fix: resolve all compilation errors in Rust/Tauri backend
- Add missing Serialize/Deserialize derives to types used in IPC
- Add missing methods to engines (CompositionEngine, TipEcosystemEngine,
  SponsorMarketplaceEngine, TelemetryEngine, PerformanceEngine)
- Fix type mismatches (Result wrapping, Option handling)
- Fix lifetime parameters in async Tauri commands
- Fix MutexGuard held across await points in analytics commands
- Fix apply_for_sponsorship to construct SponsorshipApplication struct
- Add can_undo/can_redo fields to UndoRedoInfo
- Fix engine initialization (unwrap Result types with expect)
- Fix fmt import conflict in logging.rs
- Fix plugin_metadata macro import in cli.rs
- Add From<PluginError> impl for CliError
- Fix Option<f32> vs f32 type mismatch in profiler.rs
- Fix String vs &str issues in cli.rs plugin manager calls
- Add Default derive to RealTimeAnalytics

Project now compiles successfully with cargo check (warnings only).

+++ src-tauri/src/live_captions.rs
@@ -192,7 +192,7 @@ impl LiveCaptionsEngine {
         let total_words: usize = self.segments.iter().map(|s| s.text.split_whitespace().count()).sum();
         let total_minutes = self.stats.total_duration / 60.0;
         if total_minutes > 0.0 {
-            self.stats.words_per_minute = total_words as f32 / total_minutes;
+            self.stats.words_per_minute = total_words as f32 / total_minutes as f32;
         }
         
         self.stats.active_language = Some(self.config.language.clone());
